Refactored the migrate method to support a one migration per batch mode
The reason for this is that some users of the engine have large batches and these cause deadlocks. So running one migration at a time puts the power back into the developer of the migration's hands to ensure no deadlocks occur.

Refactored the parameters taken by Run to be cmd + options... to allow for future enhancements without breaking interface changes. Included a bunch of needed documentation.